Implemented shoot by space
Implemented shoot ability using space and holding it to increase the powerslider, further tweaking may be implied for custom increased speed which actually is integer step. Also, I meant to fix the ability to shoot while the ball is in move then I noticed it was a feature all along so I skipped that.